+++ examples/free.rs
@@ -1,157 +1,37 @@
-use bevy::input::mouse::MouseMotion;
-use bevy::prelude::*;
-use bevy_dolly::*;
+mod helpers;
 
-#[derive(Component)]
-struct MainCamera;
+use bevy::prelude::*;
+use bevy_dolly::prelude::*;
+use helpers::spawn_example_scene;
 
 fn main() {
     App::new()
         .insert_resource(Msaa { samples: 4 })
         .add_plugins(DefaultPlugins)
-        .add_startup_system(setup)
-        .add_startup_system(initial_grab_cursor)
-        .add_system(update_camera_system)
-        .add_system(cursor_grab_system)
+
+        // Add Dolly plugin
+        .add_plugin(DollyPlugin)
+
+        .add_startup_system(spawn_example_scene)
+        .add_startup_system(spawn_camera)
         .run();
 }
 
 /// set up a simple 3D scene
-fn setup(
-    mut commands: Commands,
-    mut meshes: ResMut<Assets<Mesh>>,
-    mut materials: ResMut<Assets<StandardMaterial>>,
-    asset_server: Res<AssetServer>,
-) {
-    // plane
-    commands.spawn_bundle(PbrBundle {
-        mesh: meshes.add(Mesh::from(shape::Plane { size: 5.0 })),
-        material: materials.add(Color::rgb(0.3, 0.5, 0.3).into()),
+fn spawn_camera(mut commands: Commands) {
+
+    // How you can spawn a normal camera
+    commands.spawn_bundle(PerspectiveCameraBundle {
+        transform: Transform::from_xyz(0.0, 2.0, -5.0)
+        // Look at something
+            .looking_at( Vec3::ZERO, Vec3::Y),
         ..Default::default()
     });
 
-    commands
-        .spawn_bundle((
-            Transform {
-                translation: Vec3::new(0., 0.2, 0.),
-                ..Default::default()
-            },
-            GlobalTransform::identity(),
-        ))
-        .with_children(|cell| {
-            cell.spawn_scene(asset_server.load("poly_dolly.gltf#Scene0"));
-        })
-        .id();
-
-    let transform = Transform::from_xyz(-2.0f32, 2.0f32, 5.0f32).looking_at(Vec3::ZERO, Vec3::Y);
-
-    let translation = transform.translation;
-    let rotation = transform.rotation;
-    let mut yaw_pitch = YawPitch::new();
-    yaw_pitch.set_rotation_quat(rotation);
-
-    commands
-        .spawn_bundle(PerspectiveCameraBundle {
-            transform,
-            ..Default::default()
-        })
-        .insert(
-            CameraRig::builder()
-                .with(Position {
-                    position: translation,
-                })
-                .with(Rotation { rotation })
-                .with(yaw_pitch)
-                .with(Smooth::new_position_rotation(1.0, 1.0))
-                .build(),
-        )
-        .insert(MainCamera);
-
-    // light
-    commands.spawn_bundle(PointLightBundle {
-        transform: Transform::from_xyz(4.0, 8.0, 4.0),
-        ..Default::default()
-    });
-}
-
-/// Grabs/ungrabs mouse cursor
-fn toggle_grab_cursor(window: &mut Window) {
-    window.set_cursor_lock_mode(!window.cursor_locked());
-    window.set_cursor_visibility(!window.cursor_visible());
-}
-
-/// Grabs the cursor when game first starts
-fn initial_grab_cursor(mut windows: ResMut<Windows>) {
-    toggle_grab_cursor(windows.get_primary_mut().unwrap());
-}
-
-fn cursor_grab_system(keys: Res<Input<KeyCode>>, mut windows: ResMut<Windows>) {
-    let window = windows.get_primary_mut().unwrap();
-    if keys.just_pressed(KeyCode::Escape) {
-        toggle_grab_cursor(window);
-    }
-}
-
-#[allow(clippy::type_complexity)]
-fn update_camera_system(
-    time: Res<Time>,
-    keys: Res<Input<KeyCode>>,
-    windows: Res<Windows>,
-    mut mouse_motion_events: EventReader<MouseMotion>,
-    mut query: Query<(&mut Transform, &mut CameraRig)>,
-) {
-    let time_delta_seconds: f32 = time.delta_seconds();
-    let boost_mult = 5.0f32;
-    let sensitivity = Vec2::splat(1.0);
-
-    let mut move_vec = Vec3::ZERO;
-
-    // Q: Is dolly left-handed so z is flipped?
-    if keys.pressed(KeyCode::W) {
-        move_vec.z -= 1.0;
-    }
-    if keys.pressed(KeyCode::S) {
-        move_vec.z += 1.0;
-    }
-    if keys.pressed(KeyCode::A) {
-        move_vec.x -= 1.0;
-    }
-    if keys.pressed(KeyCode::D) {
-        move_vec.x += 1.0;
-    }
-
-    if keys.pressed(KeyCode::E) || keys.pressed(KeyCode::Space) {
-        move_vec.y += 1.0;
-    }
-    if keys.pressed(KeyCode::Q) {
-        move_vec.y -= 1.0;
-    }
-
-    let boost: f32 = if keys.pressed(KeyCode::LShift) {
-        1.
-    } else {
-        0.
-    };
-
-    let mut delta = Vec2::ZERO;
-    for event in mouse_motion_events.iter() {
-        delta += event.delta;
-    }
-
-    for (mut t, mut rig) in query.iter_mut() {
-        let move_vec =
-            rig.final_transform.rotation * move_vec.clamp_length_max(1.0) * boost_mult.powf(boost);
-
-        let window = windows.get_primary().unwrap();
-        if window.cursor_locked() {
-            rig.driver_mut::<YawPitch>().rotate_yaw_pitch(
-                -0.1 * delta.x * sensitivity.x,
-                -0.1 * delta.y * sensitivity.y,
-            );
-            rig.driver_mut::<Position>()
-                .translate(move_vec * time_delta_seconds * 10.0);
-        }
-
-        *t = rig.update(time_delta_seconds);
-    }
+    // Create our camera with defaults, currently that is free look
+    // commands.spawn_bundle(DollyCameraBundle {
+    //     // We can we create it anywhere we want
+    //     transform: Transform::from_xyz(0.0, 2.0, -5.0),
+    //     ..Default::default()
+    // });
 }
